requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1
django-tastypie==0.14.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-mimeparse==1.6.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
